requirements.txt
biopython==1.78
certifi==2021.5.30
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
scipy==1.7.0
six==1.16.0
